feat: move story creation to dedicated route
